win: fix invalid print format for off_t variables

+++ src/test/blk_non_zero/blk_non_zero.c
@@ -162,31 +162,33 @@ main(int argc, char *argv[])
 		switch (argv[read_arg][0]) {
 		case 'r':
 			if (pmemblk_read(handle, buf, lba) < 0)
-				OUT("!read      lba %zu", lba);
+				OUT("!read      lba %zu", (size_t)lba);
 			else
-				OUT("read      lba %zu: %s", lba, ident(buf));
+				OUT("read      lba %zu: %s", (size_t)lba,
+					ident(buf));
 			break;
 
 		case 'w':
 			construct(buf);
 			if (pmemblk_write(handle, buf, lba) < 0)
-				OUT("!write     lba %zu", lba);
+				OUT("!write     lba %zu", (size_t)lba);
 			else
-				OUT("write     lba %zu: %s", lba, ident(buf));
+				OUT("write     lba %zu: %s", (size_t)lba,
+					ident(buf));
 			break;
 
 		case 'z':
 			if (pmemblk_set_zero(handle, lba) < 0)
-				OUT("!set_zero  lba %zu", lba);
+				OUT("!set_zero  lba %zu", (size_t)lba);
 			else
-				OUT("set_zero  lba %zu", lba);
+				OUT("set_zero  lba %zu", (size_t)lba);
 			break;
 
 		case 'e':
 			if (pmemblk_set_error(handle, lba) < 0)
-				OUT("!set_error lba %zu", lba);
+				OUT("!set_error lba %zu", (size_t)lba);
 			else
-				OUT("set_error lba %zu", lba);
+				OUT("set_error lba %zu", (size_t)lba);
 			break;
 		}
 	}
